# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Edu app API",
        "description": "API documentation for Edu app.",
        "contact": {
            "email": "eamhel98@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base URL"
        }
    ],
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get all courses with optional filters",
                "description": "Returns a paginated list of courses, optionally filtered by city and/or field",
                "operationId": "getCourses",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City name to filter courses",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "field",
                        "in": "query",
                        "description": "Field name to filter courses",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total_pages": {
                                            "type": "integer"
                                        },
                                        "total_courses": {
                                            "type": "integer"
                                        },
                                        "courses_per_page": {
                                            "type": "integer"
                                        },
                                        "courses": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "integer"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "content": {
                                                        "type": "string"
                                                    },
                                                    "provider": {
                                                        "type": "string"
                                                    },
                                                    "city": {
                                                        "type": "string"
                                                    },
                                                    "field": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch courses"
                    }
                }
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Store a new course",
                "description": "Creates a new course with validated data and stores it in the database.",
                "operationId": "storeCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "provider_id",
                                    "field_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laravel Mastery"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "provider_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "field_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A complete Laravel 11 course."
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "content": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course created successfully"
                                        },
                                        "course": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Laravel Mastery"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "images/laravel.png"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "content/laravel.pdf"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Failed to create course"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a specific course by ID",
                "description": "Returns details of a specific course including provider, field, and reviews",
                "operationId": "getCourseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "course": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "500": {
                        "description": "Failed to fetch course"
                    }
                }
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update an existing course",
                "description": "Updates a course owned by the authenticated user",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Course"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 300
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "content": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully"
                    },
                    "404": {
                        "description": "Provider not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Failed to update course"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course",
                "description": "Deletes a course by ID",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "500": {
                        "description": "Failed to delete course"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}